Refactor dashboard view model into focused modules
